Se finalizó la modularización de la totalidad del código, resultando en la creación de las carpetas utils y components dentro de src las cuales alojan todas las clases y funciones utilizadas, al tiempo que se optimizó la totalidad del código
