--- src/classes/arrays/features.js
@@ -1,17 +1,0 @@
-export const arrFeatures = [
-    {
-        icon: 'bi bi-star-fill',
-        title: 'Calidad Premium',
-        description: 'Utilizamos materiales de primera calidad para garantizar la durabilidad y comodidad de nuestro calzado.'
-    },
-    {
-        icon: 'bi bi-tree-fill',
-        title: 'Materiales Sustentables',
-        description: 'Nos preocupamos por el medio ambiente utilizando materiales ecológicos y procesos sostenibles.'
-    },
-    {
-        icon: 'bi bi-heart-fill',
-        title: 'Hecho con Amor',
-        description: 'Cada par de zapatos es fabricado con dedicación y atención a los detalles.'
-    }
-];
