--- src/classes/utils/ContactForm.js
@@ -1,29 +1,0 @@
-export default class ContactForm {
-    constructor(email) {
-        this.contactForm = `
-        <form id="contact-form" action="https://formsubmit.co/${email}" method="POST">
-            <!-- Campos ocultos de FormSubmit -->
-            <input type="hidden" name="_subject" value="Nuevo mensaje desde la web">
-            <input type="hidden" name="_template" value="table">
-            <input type="hidden" name="_next" value="https://blackanvilsoftworks.github.io/CalwillYFiorella/#contact-form-container">
-            <input type="hidden" name="_captcha" value="false">
-            
-            <div class="col-12 col-md-6 mb-3 px-1">
-                <label for="name" class="form-label">Nombre</label>
-                <input type="text" name="Nombre" class="form-control" id="name" placeholder="Solo letras" required>
-            </div>
-            <div class="col-12 col-md-6 mb-3 px-1">
-                <label for="phone-number" class="form-label">Número de Teléfono</label>
-                <input type="phone-number" name="Teléfono" class="form-control" id="phone-number" placeholder="Sin espacios ni guiones 1122223333" required>
-            </div>
-            <div class="col-12 mb-3">
-                <label for="message" class="form-label">Mensaje</label>
-                <textarea class="form-control" name="Mensaje" id="message" rows="3" placeholder="Recibirá una respuesta vía WhatsApp lo más pronto posible." required></textarea>
-            </div>
-            <button type="submit" class="btn btn-primary">Enviar</button>
-        </form>
-    `;
-    }
-    
-    getContactForm() { return this.contactForm; }
-}
