+++ src/components/HeroSection.js
@@ -1,13 +1,13 @@
 export default class HeroSection {
-    constructor() {
-        this.heroSection = '';
+    constructor () {
+        // No properties needed for now
     }
 
-    getHeroSection() {
+    getHeroSection () {
         return this.generateHeroSection();
     }
 
-    generateHeroSection() {
+    generateHeroSection () {
         return `
             <div class="container">
                 <h1 class="display-4 fw-bold mb-4">Calzado de Calidad para Toda la Familia</h1>
