+++ src/components/Navbar.js
@@ -1,20 +1,21 @@
+import { globalInfo }    from '../utils/constants.js';
+import { arrContainers } from '../utils/arrays.js';
+
 export default class Navbar {
-    constructor(id, name, arrContainers) {
+    constructor (id) {
         this.id = id;
-        this.name = name;
-        this.arrContainers = arrContainers;
     }
 
-    getNavbar() {
+    getNavbar () {
         return this.generateNavbar();
     }
 
-    generateNavbar() {
+    generateNavbar () {
         return `
             <div class="container-fluid">
                 <a class="navbar-brand" href="#${this.id}">
-                    <img src="./assets/navbar-logo.png" alt="${this.name} Logo" height="40" class="d-inline-block align-text-top me-2">
-                    <span class="ms-2 navbar-title">${this.name.toUpperCase()}</span>
+                    <img src="./assets/navbar-logo.png" alt="${globalInfo.name} Logo" height="40" class="d-inline-block align-text-top me-2">
+                    <span class="ms-2 navbar-title">${globalInfo.name.toUpperCase()}</span>
                 </a>
                 <button class="navbar-toggler border-0" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
@@ -23,13 +24,13 @@ export default class Navbar {
                     <ul class="navbar-nav ms-auto">
                         ${this.createNavbarItems()}
                     </ul>
-                </div>
+                </div>S
             </div>
         `;
     }
 
-    createNavbarItems = () => {
-        return this.arrContainers.map(item => {
+    createNavbarItems () {
+        return arrContainers.map(item => {
             if (item.navbar){
                 return `
                     <li class="nav-item">
