+++ src/utils/constants.js
@@ -1,0 +1,11 @@
+export const globalInfo = {
+    name: 'calwill & fiorella',
+    phoneNumber: '+54 9 11-5959-0586',
+    email: 'calwillyfiorella@gmail.com'
+};
+
+export const imagesPath = './src/assets/images/';
+
+export const nameRegex = /^[A-Za-z\s]+$/;
+export const phoneNumberRegex = /^[0-9]{10}$/;
+export const messageRegex = /^[A-Za-z0-9\-\s]/g; // todo Que esto valide que no se metan símbolos extraños. Que se puedan poner , y .
